Address connector recovery review feedback
