+++ crates/tandem-server/src/app/state/automation/logic.rs
@@ -4,6 +4,7 @@ include!("logic_parts/part01.rs");
 include!("logic_parts/part02.rs");
 include!("logic_parts/part07.rs");
 include!("logic_parts/part03.rs");
+include!("logic_parts/connector_validation.rs");
 include!("logic_parts/part04.rs");
 include!("logic_parts/part06.rs");
 include!("logic_parts/part08.rs");
