+++ crates/tandem-server/src/app/state/automation/tests.rs
@@ -5,3 +5,4 @@ include!("tests_parts/part04.rs");
 include!("tests_parts/part05.rs");
 include!("tests_parts/part06.rs");
 include!("tests_parts/part07.rs");
+include!("tests_parts/connector_result_capture.rs");
